requirements.txt
transformers
torch
torchvision
torchaudio
fastapi
uvicorn[standard]
gunicorn
python-dotenv
gradio
